Update README.md
ABOUT US
CryptoDOMAIN.com is a wholly owned subsidiary of One Concept Media and is dedicate solely to providing it's customers with Crypto Domain Solutions in a straight forward and efficient manner.  We are an aggregator and wholesaler of these specific domains and are the largest and most complete, dedicated, marketplace for the acquisition of your Crypto Domain needs.  We offer no haggle and no auction pricing so that you can secure your domain now and begin to utilize it immediately for your needs within this ultra high-growth space.  We own every domain we sell so you know it is actually yours when you purchase. 
Contact
Customer Service:
Info@CryptoDomain.com